requirements.txt
Django==4.2.10
python-dotenv==1.0.0
oracledb==3.3.0
psycopg2-binary==2.9.9
Pillow==10.2.0